+++ Projeto/src/DSS/GestOrcamentos/IGestOrcamentosFacade.java
@@ -8,6 +8,6 @@ public interface IGestOrcamentosFacade {
     LinkedHashMap<Integer, Orcamento> obtemListaOrcamentos ();
     public Orcamento obtemOrcamentoMaisAntigo();
     public void removeOrcamentoMaisAntigo();
-    int obtemOrcamento (int nif);
+    double obtemOrcamento (int nif);
     boolean existeOrcamento (int nif);
 }
